feat: Now support multi-sided marketplace
Add a table for product owner.
The product_owner table has a constraint, product_id is unique index.
Therefore to change owners, you must perform a transaction deleting the previous owner and adding the new one.
Add a column in order_product table, to know who was the owner_id at the time the order was placed.
I added the migration for hasura, with the respective up sql and down sql.
Add the hasuna metadata	table